backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
stripe==7.8.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.6.0

# Optional dependencies
pillow==10.1.0  # For image processing
jinja2==3.1.2   # For email templates
python-slugify==8.0.1  # For URL slugs
